requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.28.0
sqlalchemy>=2.0.25,<2.1.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<1.14.0
pydantic-settings>=2.1.0,<2.3.0
jinja2>=3.1.0,<3.3.0
httpx>=0.26.0,<0.28.0
python-multipart>=0.0.9,<0.0.10
python-dotenv>=1.0.0,<2.0.0
pytest>=7.4.0,<8.0.0
psycopg2
slack-bolt>=1.18.0,<2.0.0
slack-sdk>=3.27.0,<4.0.0
python-multipart>=0.0.9,<0.0.10